+++ packages/widgetbook_generator/lib/code_generators/instances/addons/addons.dart
@@ -1,0 +1,1 @@
+export './material_theme_addon/material_theme_addon.dart';