+++ packages/widgetbook_generator/lib/generators/app_generator.dart
@@ -1,23 +1,15 @@
 import 'package:widgetbook_annotation/widgetbook_annotation.dart';
+import 'package:widgetbook_generator/code_generators/instances/addons/addon_instance.dart';
+import 'package:widgetbook_generator/code_generators/instances/addons/material_theme_addon/material_theme_addon_instance.dart';
 import 'package:widgetbook_generator/code_generators/instances/app_info_instance.dart';
-import 'package:widgetbook_generator/code_generators/instances/device_instance.dart';
-import 'package:widgetbook_generator/code_generators/instances/double_instance.dart';
-import 'package:widgetbook_generator/code_generators/instances/frame_instance.dart';
-import 'package:widgetbook_generator/code_generators/instances/function_call_instance.dart';
-import 'package:widgetbook_generator/code_generators/instances/theme_instance.dart';
 import 'package:widgetbook_generator/code_generators/instances/variable_instance.dart';
 import 'package:widgetbook_generator/code_generators/instances/widgetbook_category_instance.dart';
 import 'package:widgetbook_generator/code_generators/instances/widgetbook_instance.dart';
 import 'package:widgetbook_generator/models/widgetbook_app_builder_data.dart';
-import 'package:widgetbook_generator/models/widgetbook_device_frame_data.dart';
 import 'package:widgetbook_generator/models/widgetbook_locales_data.dart';
-import 'package:widgetbook_generator/models/widgetbook_localization_builder_data.dart';
 import 'package:widgetbook_generator/models/widgetbook_localizations_delegates_data.dart';
-import 'package:widgetbook_generator/models/widgetbook_scaffold_builder_data.dart';
-import 'package:widgetbook_generator/models/widgetbook_theme_builder_data.dart';
 import 'package:widgetbook_generator/models/widgetbook_theme_data.dart';
 import 'package:widgetbook_generator/models/widgetbook_theme_type_data.dart';
-import 'package:widgetbook_generator/models/widgetbook_use_case_builder_data.dart';
 import 'package:widgetbook_generator/models/widgetbook_use_case_data.dart';
 import 'package:widgetbook_generator/services/tree_service.dart';
 
@@ -36,49 +28,26 @@ String generateWidgetbook({
   WidgetbookThemeData? widgetbookThemeData,
   WidgetbookThemeTypeData? themeTypeData,
   required List<WidgetbookThemeData> themes,
-  WidgetbookDeviceFrameData? deviceFrameBuilder,
-  WidgetbookLocalizationBuilderData? localizationBuilder,
   WidgetbookAppBuilderData? appBuilder,
-  WidgetbookScaffoldBuilderData? scaffoldBuilder,
-  WidgetbookThemeBuilderData? themeBuilder,
-  WidgetbookUseCaseBuilderData? useCaseBuilder,
 }) {
   final category =
       _generateCategoryInstance(useCases, foldersExpanded, widgetsExpanded);
+
+  final addons = <AddOnInstance>[];
+  if (constructor == WidgetbookConstructor.material) {
+    addons.add(MaterialThemeAddonInstance(themes: themes));
+  }
+
   final widgetbookInstanceCode = WidgetbookInstance(
     constructor: constructor,
+    addons: addons,
     appInfoInstance: AppInfoInstance(name: name),
-    themes: themes.map((theme) => ThemeInstance(theme: theme)).toList(),
-    devices: devices.map((device) => DeviceInstance(device: device)).toList(),
-    frames: frames.map((frame) => FrameInstance(frame: frame)).toList(),
-    textScaleFactors: textScaleFactors.map(DoubleInstance.value).toList(),
     categories: [
       category,
     ],
     type: themeTypeData?.name,
-    locales: localesData != null
-        ? VariableInstance(variableIdentifier: localesData.name)
-        : null,
-    localizationDelegates: localizationDelegatesData != null
-        ? VariableInstance(variableIdentifier: localizationDelegatesData.name)
-        : null,
-    deviceFrameBuilder: deviceFrameBuilder != null
-        ? VariableInstance(variableIdentifier: deviceFrameBuilder.name)
-        : null,
-    localizationBuilder: localizationBuilder != null
-        ? VariableInstance(variableIdentifier: localizationBuilder.name)
-        : null,
     appBuilder: appBuilder != null
         ? VariableInstance(variableIdentifier: appBuilder.name)
-        : null,
-    scaffoldBuilder: scaffoldBuilder != null
-        ? VariableInstance(variableIdentifier: scaffoldBuilder.name)
-        : null,
-    themeBuilder: themeBuilder != null
-        ? FunctionCallInstance(name: themeBuilder.name)
-        : null,
-    useCaseBuilder: useCaseBuilder != null
-        ? VariableInstance(variableIdentifier: useCaseBuilder.name)
         : null,
   ).toCode();