+++ packages/widgetbook_generator/test/src/code_generators/instances/addons/material_theme_addon/material_theme_addon_instance_test.dart
@@ -1,0 +1,17 @@
+import 'package:test/expect.dart';
+import 'package:test/scaffolding.dart';
+import 'package:widgetbook_generator/code_generators/instances/instances.dart';
+
+import '../../../../mocks/mocks.dart';
+
+void main() {
+  group(
+    '$MaterialThemeAddonInstance',
+    () {
+      test('.name is "MaterialThemeAddon"', () {
+        final sut = MaterialThemeAddonInstance(themes: themes);
+        expect(sut.name, equals('MaterialThemeAddon'));
+      });
+    },
+  );
+}